refactor out constant country translations
